# api-doc.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/user-login":{"post":{"tags":["user-controller"],"summary":"Exchange Google OIDC Token for JWT","operationId":"exchange","parameters":[{"name":"OIDC-Token","in":"header","description":"Google OIDC Token","required":true}],"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}},"401":{"description":"Invalid token","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"example":"{\n      \"title\": \"Token validation\",\n      \"status\": 401,\n      \"detail\": \"Invalid token\",\n}\n"}}}}}},"/tokens/refresh":{"post":{"tags":["token-controller"],"summary":"Refresh new access token","operationId":"refresh","parameters":[{"name":"refresh_token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"New access token is generated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}},"401":{"description":"Expired token","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"example":{"title":"Token validation","status":401,"detail":"Expired token","instance":"/tokens/refresh"}}}}}}},"/profiles/me/password":{"post":{"tags":["profile-controller"],"summary":"Change password","description":"Intended for admin","operationId":"changePassword","parameters":[{"name":"password","in":"query","required":false,"schema":{"type":"string"}},{"name":"new_password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"403":{"description":"The user is not an admin"},"400":{"description":"Request validation or Wrong confirm password","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"examples":{"Request validation":{"$ref":"#/components/examples/Validation"},"Password mismatched":{"description":"Password mismatched","value":{"type":"https://example.com/errors/password-mismatch","title":"Confirm password constraint","status":400,"detail":"Wrong password","instance":"/profiles/me/password"}}}}}},"200":{"description":"Password is changed, returning profile","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/files/upload":{"post":{"tags":["presign-controller"],"summary":"Generate presigned url for uploading file","operationId":"uploadUrl","parameters":[{"name":"filename","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/S3PresignedResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/docs":{"get":{"tags":["document-controller"],"description":"Search documents","operationId":"search","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","minLength":1}},{"name":"start","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"end","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"type","in":"query","required":true,"schema":{"type":"string","enum":["PDF","DOCX"]}},{"name":"lastDocId","in":"query","required":false,"schema":{"type":"string"}},{"name":"lastDocScore","in":"query","required":false,"schema":{"type":"number","format":"float"}}],"responses":{"200":{"description":"The documents","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocPage"}}}},"400":{"description":"Validation failure","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"example":null}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["document-controller"],"description":"Add a new document","operationId":"add","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDocumentRequest"}}},"required":true},"responses":{"202":{"description":"Document added successfully"},"400":{"description":"Validation failure","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"examples":{"Request validation":{"$ref":"#/components/examples/Validation"},"File integrity violated":{"description":"File integrity violated","value":{"title":"File integrity error","status":400,"detail":"Mismatch Etag","instance":"/tokens/refresh"}}}}}}},"security":[{"bearerAuth":[]}]}},"/docs/suggest":{"post":{"tags":["document-controller"],"description":"Suggest documents","operationId":"suggest","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The suggestion text stream","content":{"text/plain":{"example":"      Hello, this is teri aka decade\n"}}}},"security":[{"bearerAuth":[]}]}},"/admins":{"post":{"tags":["user-controller"],"summary":"Register new admin","description":"Require an admin authentication","operationId":"registerAdmin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequest"}}},"required":true},"responses":{"201":{"description":"New admin is created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}},"403":{"description":"Require an admin authentication"},"400":{"description":"Validation","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"examples":{"Request validation":{"$ref":"#/components/examples/Validation"}}}}},"409":{"description":"Username constraint","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"examples":{"User already exist":{"description":"User already exist","value":{"title":"Username constraint","status":409,"detail":"User already exist","instance":"/profiles/me/password"}}}}}}},"security":[{"bearerAuth":[]}]}},"/profiles/me":{"get":{"tags":["profile-controller"],"description":"Get user profile information","operationId":"get","responses":{"200":{"description":"User profile information","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}},"security":[{"bearerAuth":[]}]},"patch":{"tags":["profile-controller"],"description":"Change user profile","operationId":"changeProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileRequest"}}},"required":true},"responses":{"200":{"description":"User profile after update","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/docs/{id}":{"get":{"tags":["document-controller"],"description":"Find a document by id","operationId":"find","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"The document","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentResponse"}}}},"404":{"description":"Document not found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"example":null}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{}}}},"AccessToken":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"AccountResponse":{"type":"object","properties":{"profile":{"$ref":"#/components/schemas/ProfileResponse"},"accessToken":{"$ref":"#/components/schemas/AccessToken"}}},"ProfileResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"name":{"type":"string"},"dob":{"type":"string","format":"date-time"},"role":{"type":"string"},"gender":{"type":"string"}}},"S3PresignedResponse":{"type":"object","properties":{"fileKey":{"type":"string"},"presignedUploadUrl":{"type":"string"}}},"CreateDocumentRequest":{"type":"object","properties":{"filename":{"type":"string","minLength":1},"eTag":{"type":"string"},"fileKey":{"type":"string"},"title":{"type":"string","minLength":1},"description":{"type":"string","minLength":1},"type":{"type":"string","enum":["PDF","DOCX"]}},"required":["description","eTag","fileKey","filename","title"]},"SignUpRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":20,"minLength":5,"pattern":"\\S+"},"password":{"type":"string"},"name":{"type":"string","minLength":1},"gender":{"type":"number","format":"float"},"dob":{"type":"string","format":"date-time"},"avatar":{"type":"string"}},"required":["gender","name","username"]},"ProfileRequest":{"type":"object","properties":{"name":{"type":"string","pattern":"\\S.*"},"gender":{"type":"number","format":"float"},"dob":{"type":"string","format":"date-time"}}},"DocPage":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int64"},"docs":{"type":"array","items":{"$ref":"#/components/schemas/DocumentItemResponse"}}}},"DocumentItemResponse":{"type":"object","properties":{"id":{"type":"string"},"filename":{"type":"string"},"title":{"type":"string"},"fileType":{"type":"string","enum":["PDF","DOCX"]},"createdAt":{"type":"string","format":"date-time"},"score":{"type":"number","format":"float"}}},"DocumentResponse":{"type":"object","properties":{"id":{"type":"string"},"filename":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"fileType":{"type":"string","enum":["PDF","DOCX"]},"createdAt":{"type":"string","format":"date-time"},"score":{"type":"number","format":"float"}}}},"examples":{"Validation":{"summary":"Validation failure","value":{"status":400,"type":"https://example.com/errors/validation","detail":"Email is invalid","title":"Validation failure"}},"NotFound":{"summary":"Resource not found","value":{"status":404,"title":"The requested resource was not found"}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"},"basicAuth":{"type":"http","scheme":"basic"}}}}